% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gomoku_board.R
\name{gomoku_board}
\alias{gomoku_board}
\title{Create a new gomoku_board}
\usage{
gomoku_board(board_size = 19)
}
\arguments{
\item{board_size}{The dimensions of the board, to be n x n where n = board_size. Defaults to the standard Gomoku board size of 19x19, but also accepts 15 as an alternative.}
}
\value{
A ggplot2 object of a board corresponding to the size specified by the user.
}
\description{
gomoku_board() draws a 19x19 (default) or 15x15 board for playing the board game Gomoku.
}
\examples{
# Both of these function generate a 19x19 Gomoku board
gomoku_board(19)
gomoku_board()

# Generates a 15x15 Gomoku board
gomoku_board(15)
}
\author{
Pin-Chun Chen and J Steven Raquel
}
