% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gomoku_input.R
\name{gomoku_input}
\alias{gomoku_input}
\title{Checking valid user input for gomoku_play}
\usage{
gomoku_input(board_size)
}
\arguments{
\item{board_size}{a numeric variable of either 19 (default) or 15, indicating whether to restrict X and Y coordinates inputs for a 19x19 (default) or 15x15 board for playing the board game}
}
\value{
One of the integer coordinates between 1 and board_size to be used as either an x or y coordinate on the gomoku board.
}
\description{
This function is designed not to be called directly but to be used within the gomoku_play() function in which it prompts the user for a valid input.
}
\examples{
# When called during a game that used gomoku_board() or gomoku_board(19)
gomoku_input(board_size = 19)
# When called during agame that used gomoku_board(15)
gomoku_input(board_size = 15)
}
\author{
Pin-Chun Chen and J Steven Raquel
}
