% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gomoku_play.R
\name{gomoku_play}
\alias{gomoku_play}
\title{Play Gomoku}
\usage{
gomoku_play(board, show_moves = FALSE)
}
\arguments{
\item{board}{the output of the function "gomoku_board", which is a data frame of the ggplot gomoku board}

\item{show_moves}{a boolean variable (TRUE or FALSE) indicating whether to show the current number of moves on the tiles}
}
\description{
gomoku_play() starts a game of standard Gomoku using the object generated by gomoku_board().
}
\examples{
# Starts a game of Gomoku that does not show move numbers on tiles
gomoku_play(gomoku_board())

# Starts a game of Gomoku that does show move numbers on tiles
gomoku_play(gomoku_board(), show_moves = TRUE)
}
\author{
Pin-Chun Chen and J Steven Raquel
}
