% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gomoku_victory.R
\name{gomoku_victory}
\alias{gomoku_victory}
\title{Checking for a winner of the Gomoku game}
\usage{
gomoku_victory(matrix)
}
\arguments{
\item{matrix}{This input matrix is the matrix completed by the user input given by gomoku_input(), which is itself called by gomoku_play(). It mirrors exactly the board drawn by ggplot2.}
}
\value{
A string containing either "black" or "white", depending on who won, once a winner is found.
}
\description{
This function is designed not to be called directly but to be used within the gomoku_play() function where it maps the nested function gomoku_winner() to all rows, columns and diagonals of the game board in order to detect a winner.
}
\author{
Pin-Chun Chen and J Steven Raquel
}
