% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gomoku_winner.R
\name{gomoku_winner}
\alias{gomoku_winner}
\title{Detects the name of the player who won the game of Gomoku}
\usage{
gomoku_winner(rle_output)
}
\arguments{
\item{rle_output}{The list generated by the lapply(x, FUN =  rle) call within the gomoku_victory() function, where x is a vector containing the indices of the rows, columns or diagonals respectively.}
}
\value{
a list that indicate whether black or white is the winner based on the output of the rle function
}
\description{
This function is not called directly. It is called within gomoku_victory() and is designed to be mapped to the rows, columns, and diagonals of the matrix generated by gomoku_play(). Using the rle() function it seeks out an instance in which either player has gotten 5 of their own tiles in a row, then stores this player's name (either "black" or "white") into a variable which is then returned.
}
\author{
Pin-Chun Chen and J Steven Raquel
}
